fixing ci bugs once again
